requirements.txt
datasets~=2.10.0
fortepyan~=0.2.4
hydra-core~=1.3.2
matplotlib~=3.7.2
numpy~=1.25.2
pandas~=1.5.3
scikit-learn~=1.3.0
streamlit~=1.25.0
torch~=2.0.1
transformers~=4.31.0
